move functions into separate misc module
