+++ gpflow/utilities/utilities.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2020 The GPflow Contributors. All Rights Reserved.
+# Copyright 2017-2021 The GPflow Contributors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,22 +15,33 @@
 import copy
 import re
 from functools import lru_cache
-from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, TypeVar, Union
+from typing import Any, Callable, Dict, Optional, Tuple, TypeVar, Union
 
 import numpy as np
 import tensorflow as tf
 import tensorflow_probability as tfp
+from deprecated import deprecated
 from packaging.version import Version
 from tabulate import tabulate
 
 from ..base import Parameter
-from ..config import default_float, default_int, default_summary_fmt
-from .ops import cast
+from ..config import default_summary_fmt
+from .misc import (  # for backwards-compatibility
+    set_trainable,
+    to_default_float,
+    to_default_int,
+    training_loop,
+)
+
+to_default_float = deprecated(reason="moved to gpflow.utilities.misc.to_default_float")(
+    to_default_float
+)
+to_default_int = deprecated(reason="moved to gpflow.utilities.misc.to_default_int")(to_default_int)
+training_loop = deprecated(reason="moved to gpflow.utilities.misc.training_loop")(training_loop)
+set_trainable = deprecated(reason="moved to gpflow.utilities.misc.set_trainable")(set_trainable)
 
 __all__ = [
-    "set_trainable",
     "multiple_assign",
-    "training_loop",
     "print_summary",
     "tabulate_module_summary",
     "deepcopy",
@@ -38,10 +49,13 @@ __all__ = [
     "leaf_components",
     "parameter_dict",
     "read_values",
-    "to_default_float",
-    "to_default_int",
     "reset_cache_bijectors",
     "select_dict_parameters_with_prior",
+    # for backwards-compatibility:
+    "to_default_float",
+    "to_default_int",
+    "set_trainable",
+    "training_loop",
 ]
 
 TraverseInput = TypeVar("TraverseInput", tf.Variable, tf.Module, Parameter)
@@ -49,26 +63,6 @@ State = Any
 Path = str
 Accumulator = Tuple[Path, State]
 TraverseUpdateCallable = Callable[[TraverseInput, Path, State], State]
-
-
-def to_default_int(x):
-    return cast(x, dtype=default_int())
-
-
-def to_default_float(x):
-    return cast(x, dtype=default_float())
-
-
-def set_trainable(model: Union[tf.Module, Iterable[tf.Module]], flag: bool) -> None:
-    """
-    Set trainable flag for all `tf.Variable`s and `gpflow.Parameter`s in a `tf.Module` or collection
-    of `tf.Module`s.
-    """
-    modules = [model] if isinstance(model, tf.Module) else model
-
-    for mod in modules:
-        for variable in mod.variables:
-            variable._trainable = flag
 
 
 def multiple_assign(module: tf.Module, parameters: Dict[str, tf.Tensor]):
@@ -112,41 +106,6 @@ def parameter_dict(module: tf.Module) -> Dict[str, Union[Parameter, tf.Variable]
     """
     param_dict = leaf_components(module)
     return {f".{key.split('.', 1)[-1]}": value for key, value in param_dict.items()}
-
-
-def training_loop(
-    closure: Callable[[], tf.Tensor],
-    optimizer: Optional[tf.optimizers.Optimizer] = None,
-    var_list: List[tf.Variable] = None,
-    maxiter=1e3,
-    compile=False,
-):
-    """
-    Simple generic training loop. At each iteration uses a GradientTape to compute
-    the gradients of a loss function with respect to a set of variables.
-
-    :param closure: Callable that constructs a loss function based on data and model being trained
-    :param optimizer: tf.optimizers or tf.keras.optimizers that updates variables by applying the
-        corresponding loss gradients. Adam is a default optimizer with default settings.
-    :param var_list: List of model variables to be learnt during training
-    :param maxiter: Maximum number of
-    :return:
-    """
-
-    optimizer = tf.optimizers.Adam() if optimizer is None else optimizer
-
-    def optimization_step():
-        with tf.GradientTape(watch_accessed_variables=False) as tape:
-            tape.watch(var_list)
-            loss = closure()
-        grads = tape.gradient(loss, var_list)
-        optimizer.apply_gradients(zip(grads, var_list))
-
-    if compile:
-        optimization_step = tf.function(optimization_step)
-
-    for _ in range(int(maxiter)):
-        optimization_step()
 
 
 def print_summary(module: tf.Module, fmt: str = None):
